Move count types to seperate file
